Add connection timeout to Custodia client
Custodia client classes have grown a default timeout instance attribute.
The forwarder plugin and CLI have a new timeout option. The default
timeout for both is 10 seconds.

Closes: #208

+++ tests/test_custodia.py
@@ -242,6 +242,8 @@ class CustodiaTests(unittest.TestCase):
 
     @classmethod
     def tearDownClass(cls):
+        if cls.custodia_process.poll() is not None:
+            raise AssertionError("Custodia server crashed, see test_log.txt")
         cls.custodia_process.terminate()
         cls.custodia_process.wait()
